We derive the BPLMOD value from the surface bytes per row value and dynamically adjust it based on interlace vs non-interlace display mode

+++ Kernel/Sources/driver/amiga/graphics/Surface.h
@@ -31,8 +31,8 @@ enum {
 typedef struct Surface {
     uint8_t* _Nullable      plane[MAX_PLANE_COUNT];
     CLUTEntry* _Nullable    clut;
-    int                     width;
-    int                     height;
+    size_t                  width;
+    size_t                  height;
     size_t                  bytesPerRow;
     size_t                  bytesPerPlane;
     int16_t                 clutEntryCount;
@@ -42,7 +42,7 @@ typedef struct Surface {
 } Surface;
 
 
-extern errno_t Surface_Create(int width, int height, PixelFormat pixelFormat, Surface* _Nullable * _Nonnull pOutSelf);
+extern errno_t Surface_Create(size_t width, size_t height, PixelFormat pixelFormat, Surface* _Nullable * _Nonnull pOutSelf);
 extern void Surface_Destroy(Surface* _Nullable self);
 
 // Returns the pixel width of the surface.
